+++ src/com/apmath/loans/application/v1/mappers/LoansMapper.kt
@@ -1,6 +1,6 @@
 package com.apmath.loans.application.v1.mappers
 
-import com.apmath.loans.application.v1.models.out.ResponseLoan
+import com.apmath.loans.application.v1.models.outgoing.ResponseLoan
 import com.apmath.loans.domain.models.loans.LoanInterface
 
 class LoansMapper {
